refactor: step1 피드백 적용 - Numbers -> PositiveNumbers 변경
